prepare release action: Make version input optional
The idea is to run the action twice. The first time without a version
input. It will generate a CHANGELOG.md update and raises a PR. Then the
changelog can be reviewed. based on the categories, the correct new
version can be determined and the action can be executed again.